StreamOperations: let skip() take long, not int - wider case
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@13257 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Extra/org/objectweb/proactive/extra/vfsprovider/protocol/StreamOperations.java
@@ -151,7 +151,7 @@ public interface StreamOperations {
      * @throws WrongStreamTypeException
      *             when mode of a stream does not allow to call this method
      */
-    public abstract long streamSkip(long stream, int bytes) throws IOException, StreamNotFoundException,
+    public abstract long streamSkip(long stream, long bytes) throws IOException, StreamNotFoundException,
             WrongStreamTypeException;
 
     /**
